Move almost all the buggy Orchard key code into a test-only module
