+++ zebra-chain/src/orchard/arbitrary/keys/tests.rs
@@ -1,0 +1,5 @@
+//! Tests for Orchard keys and addresses.
+
+mod generate;
+mod prop;
+mod vectors;
